Found and (hopefully) fixed bug where substantial portions of text were ignored.
All the examples on the Floki hexdocs page show the text alone in a list at the end of the HTML node tuple (i.e., they're in the form {tag, [attributes], [text] ). Since I wasn't thinking much about HTML's format when writing the external_site functions or test, I assumed that all text must therefore be in an empty list - and thus ignored any text that had a sibling HTML node.

Of course, HTML trees are actually in the form of a list of nodes where nodes are either plain text or an HTML tag represented as a tuple of {tag, [attributes], [child-nodes]} - e.g. "<i>one</i>example" becomes [{"i", [], ["one"]}, "example"]. Previously, 'example' was ignored - this update is so that it will also be transformed.