Implemented some more basic translation behaviour so tests from commit [26d81a0e6f] now pass (but many more are needed).
Meaningful capitalisations and non-word characters are now used to mark up the replacement word. Additionally, the _ character and digits are now ignored when comparing the user-specified vocab word to words contained in the HTML page. Ignoring digits may require some further consideration; i.e., it may only make sense to ignore them if they begin or end a word (e.g., word1 and 1word should both be translated, but maybe wo1rd should not). I don't think this will matter in practice - but perhaps it could be a problem on scientific / mathematical sites.