Added the ability to click on translated words and see details about them.
This commit adds a div that can be shown or hidden by clicking the word (using javascript and terrible css) that contains the original word and any additional information provided. Tests still need to be written, the audio link functionality needs to be implemented properly and the css needs significant improvement.

Additionally, this commit contains a bug fix for when comments are included in the HTML being parsed - Floki returns a list in this case, which previously caused an error since external_site.ex had no matching 'mapped_html' function.