Updates so that the translated page is embedded better into the site and so that the translated page can be fetched using any HTTP method.
Using an iframe might be a more natural way to embed the translated page, but I'm not entirely sure how to achieve this while still translating words. The srcdoc attribute could be used for the first page but then if the user clicks on a link, that will either take them to an untranslated page or, if all the links are updated, will create another nested iframe inside the first one.